basic functionality done

+++ src/Components/DashboardBody/DashboardBody.js
@@ -5,28 +5,55 @@ import { Button, Table } from 'react-bootstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faEdit, faTrash } from '@fortawesome/free-solid-svg-icons';
 import { Link } from 'react-router-dom';
+import { useState } from 'react';
 
 const DashboardBody = () => {
     const [ users, setUsers ] = useUsers();
+    const [ success, setSuccess ] = useState( false );
+
+    //Delete a user from API Server. Since the server is with fake data, I just put the code without testing. I have implemented the delete function in the local storage database.
+    // const handleDeleteUser = id => {
+    //     const proceed = window.confirm( "Are you sure to delete the booking data?" );
+    //     if ( proceed ) {
+    //         const url = `https://my-json-server.typicode.com/karolkproexe/jsonplaceholderdb/data/${ id }`;
+    //         fetch( url, {
+    //             method: 'DELETE'
+    //         } )
+    //             .then( res => res.json() )
+    //             .then( data => {
+    //                 if ( data.deletedCount > 0 ) {
+    //                     const remainingUsers = users.filter( user => parseInt( order.id ) !== parseInt( id ) );
+    //                     setUsers( remainingUsers );
+    //                     alert( 'User Data Deleted Successfully' );
+    //                     window.location.reload( false );
+    //                 }
+    //             } )
+    //     }
+    // }
 
     //DELETE a User Data from local storage
     const handleDeleteUser = id => {
-        const usersFromLocalStorage = JSON.parse( localStorage.getItem( 'users' ) );
-        const filteredUsers = usersFromLocalStorage.filter( user => user.id !== id );
-        setUsers( filteredUsers );
-        localStorage.setItem( 'users', JSON.stringify( filteredUsers ) );
+        const proceed = window.confirm( "Are you sure to delete the user data?" );
+
+        if ( proceed ) {
+            const usersFromLocalStorage = JSON.parse( localStorage.getItem( 'users' ) );
+            const filteredUsers = usersFromLocalStorage.filter( user => user.id !== id );
+            setUsers( filteredUsers );
+            localStorage.setItem( 'users', JSON.stringify( filteredUsers ) );
+            setSuccess( true );
+        }
     }
 
     return (
         <Container className="my-5 text-start">
             <h2 className="my-5">Dashboard</h2>
+            { success && <div className="alert alert-success">User Deleted Successfully</div> }
             <div>
                 <hr />
                 <div className="d-flex justify-content-between mb-3">
                     <h2>User List</h2>
                     <Link to="/users/addUser"><Button className="btn btn-primary">Add New User</Button></Link>
                 </div>
-
                 <Table striped bordered responsive>
                     <thead>
                         <tr>
